Futher split off JsonPrettyTable into own module (- WIP #316 -)
