+++ pip-licenses/output/JsonPrettyTable.py
@@ -1,0 +1,65 @@
+#!/usr/bin/env python
+# vim:fenc=utf-8 ff=unix ft=python ts=4 sw=4 sts=4 si et
+"""
+pip-licenses.output
+
+MIT License
+
+Copyright (c) 2018-2025 raimon
+Copyright (c) 2025-2026 Mr. Walls
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+
+from .. import (
+    __pkgname__,
+    __version__,
+)
+
+
+from prettytable import (
+    HRuleStyle,
+    PrettyTable,
+    RowType,
+)
+
+
+# import included here in order to limit dependencies
+# if not interested in JSON output,
+# then the dependency is not required
+import json
+
+
+class JsonPrettyTable(PrettyTable):
+    """PrettyTable-like class exporting to JSON"""
+
+    def format_row(self, row: RowType) -> dict[str, str | list[str]]:
+        return dict(zip(self._field_names, row))
+
+    def get_string(self, **kwargs: str | list[str]) -> str:
+        options = self._get_options(kwargs)
+        rows = self._get_rows(options)
+        lines = [self.format_row(row) for row in rows]
+        return json.dumps(lines, indent=2, sort_keys=True)
+
+
+__all__ = [
+    """JsonPrettyTable""",
+]
